backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=2.0.40",
]

[project.scripts]                                           
app = "app.main:app"                                        
